refactor: 오류 수정 및 코드리팩토링 version=minor

+++ data/src/main/java/com/petid/data/dto/request/HospitalOrderRequest.kt
@@ -3,7 +3,7 @@ package com.petid.data.dto.request
 import com.petid.domain.entity.HospitalOrderEntity
 
 data class HospitalOrderRequest(
-    val hospitalId: Int,
+    val hospitalId: Long,
     val date: String
 )
