+++ data/src/main/java/com/petid/data/dto/response/HospitalOrderResponse.kt
@@ -3,7 +3,7 @@ package com.petid.data.dto.response
 import com.petid.domain.entity.HospitalOrderEntity
 
 data class HospitalOrderResponse(
-    val hospitalId: Int,
+    val hospitalId: Long,
     val date: String
 )
