+++ domain/src/main/java/com/petid/domain/entity/HospitalOrderEntity.kt
@@ -1,6 +1,6 @@
 package com.petid.domain.entity
 
 data class HospitalOrderEntity(
-    val hospitalId: Int,
+    val hospitalId: Long,
     val date: String
 )
